+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 2026-04-22 : 0.4.0-rc.2  
 Update: Testing,preparation in enraiged core for initial release
 
+- Added email attribute to the profiles model fillable
 - Corrected issue with missing country_id when creating user
 
 ---